+++ packages/afformative/src/constants.js
@@ -11,12 +11,3 @@ export const SUGGESTIONS = {
   primitive: "primitive",
   verbose: "verbose",
 }
-
-/**
- * Assign a function to a value under this key to override the behaviour of all formatters.
- * This is mainly useful if your components supply some of their own data which should have
- * special treatment when being rendered.
- *
- * @example Formatter.format({ [FORMATTER_OVERRIDE]: () => 'N/A' })
- */
-export const FORMATTER_OVERRIDE = "@afformative/FORMATTER_OVERRIDE"