+++ packages/afformative/src/constants.test.js
@@ -1,4 +1,4 @@
-import { SUGGESTIONS, FORMATTER_OVERRIDE } from "./constants"
+import { SUGGESTIONS } from "./constants"
 
 describe("suggestions object", () => {
   it("is exported properly", () => {
@@ -7,11 +7,5 @@ describe("suggestions object", () => {
 
   it("has its values equal to its keys", () => {
     Object.entries(SUGGESTIONS).forEach(([key, value]) => expect(key).toEqual(value))
-  })
-})
-
-describe("formatter override key", () => {
-  it("is exported properly", () => {
-    expect(typeof FORMATTER_OVERRIDE).toBe("string")
   })
 })